Masses: solarsys.ssc, dwarfplanets.ssc, and outersys.ssc
Updated mass values of objects in the following files:
- solarsys.ssc, updating all satellites except Phobos and Deimos.
- dwarfplanets.ssc / outersys.ssc, wherever a mass value is available for the object, as well as updating Pluto's mass value. Note that Dysnomia's mass is obtained assuming that there is a genuine detection.

Significant figures are set to combination of value of 'Earth mass' used by Celestia, 5.972E24 kg (credit to pedroJ on Discord server), and the mass value of the object itself. The sigfig is set high to avoid spurious digits in the program.