Minor corrections to radii/diameters of comets (#235)
Thanks to @AstroChara for spotting some erroneous values inputted on the
radii of some comets. Counter-checked the values with those of JPL-SBDB
and other journals wherever available.
- Both 2P/Encke and 19P/Borrelly are 4.8 km in diameter. Previously
written incorrectly as ```# D = 4.2 km```
- C/1983 H1 is 9.2 km in diameter. Therefore, the correct radius value
is 4.6 km, not 4.2 km
- Hale-Bopp's diameter was accidentally set equal to its radius (30 km)
when its real value is actually ```# D = 60 km```

This acts as a patch update for PR #229

====

This PR also adds true colours from TCT for some of the comets, and fixed URL of Comet Arend-Roland.
- Chara