nearstars.stc update (2025-05-25) (#230)
Major data update for nearstars.stc.

This pull request lays foundation for expanding the range of nearstars.stc from 25 ly to 10 pc. Updates are concentrated on Radius and Temperature values from empirical measurements, as well as RotationPeriod, of bright stars, as well as all stars with distances between the Solar System and Epsilon Eridani.
- For stars beyond 25 light-years, solitary bright stars (as well as Gamma Leporis / AK Leporis) have been added, as well as placeholder tags for multiple bright stars with known orbits.
- Rotational elements are incomplete thus far; some stars have measured Inclination, but without AscendingNode; these entries have Inclination and AscendingNode field in, both set to 0.