Add 40 pc white dwarf catalog (#242)
Includes 1065 white dwarfs and 94 companion stars (one of which is
itself a WD) from the sample of [O'Brien et al. (2024), MNRAS 527, p.
8687-8705](https://cdsarc.cds.unistra.fr/viz-bin/cat/J/MNRAS/527/8687).
The remaining 13 WDs are already in other data files (extrasolar.stc and
nearstars.stc). I've also removed several entries from revised.stc which
correspond to objects in this catalog.

Not all WDs within 40 pc (~130 ly) are here, but completeness is almost
reached, with 28 objects missing from the source catalog, most of which
are in binary systems.