Added comment indicating the type of name
Currently the convention is for traditional star names to be applied on the entire system (unless the system already has multiple names assigned to each star), while names that the WGSN contrived (whether novel names from public/community campaigns, or asterism names applied on a star, or names inspired by culture) are to be applied to individual stars as WGSN intended. These comments should help when searching for them and for maintenance.